@@ -53,54 +53,54 @@
   },
   "devDependencies": {
     "@after-work.js/aw": "6.0.14",
-    "@babel/cli": "^7.22.9",
-    "@babel/core": "^7.22.9",
-    "@babel/eslint-parser": "7.22.9",
-    "@babel/plugin-transform-react-jsx": "^7.22.5",
-    "@babel/preset-env": "^7.22.9",
-    "@babel/preset-react": "^7.22.5",
-    "@commitlint/cli": "17.6.6",
-    "@commitlint/config-conventional": "17.6.6",
+    "@babel/cli": "^7.23.0",
+    "@babel/core": "^7.23.2",
+    "@babel/eslint-parser": "7.22.15",
+    "@babel/plugin-transform-react-jsx": "^7.22.15",
+    "@babel/preset-env": "^7.23.2",
+    "@babel/preset-react": "^7.22.15",
+    "@commitlint/cli": "17.8.1",
+    "@commitlint/config-conventional": "17.8.1",
     "@material-ui/core": "4.12.4",
     "@material-ui/styles": "4.11.5",
-    "@nebula.js/cli": "4.2.3",
-    "@nebula.js/cli-build": "4.2.3",
-    "@nebula.js/cli-sense": "4.2.3",
-    "@nebula.js/cli-serve": "4.2.3",
-    "@playwright/test": "1.36.1",
+    "@nebula.js/cli": "4.6.1",
+    "@nebula.js/cli-build": "4.6.1",
+    "@nebula.js/cli-sense": "4.6.1",
+    "@nebula.js/cli-serve": "4.6.1",
+    "@playwright/test": "1.39.0",
     "babel-plugin-istanbul": "^6.1.1",
     "d3-color": "3.1.0",
     "d3-scale": "4.0.2",
-    "eslint": "8.45.0",
+    "eslint": "8.52.0",
     "eslint-config-airbnb": "19.0.4",
-    "eslint-config-prettier": "8.8.0",
-    "eslint-plugin-import": "2.27.5",
+    "eslint-config-prettier": "8.10.0",
+    "eslint-plugin-import": "2.29.0",
     "eslint-plugin-jsx-a11y": "6.7.1",
-    "eslint-plugin-mocha": "10.1.0",
-    "eslint-plugin-playwright": "0.15.3",
+    "eslint-plugin-mocha": "10.2.0",
+    "eslint-plugin-playwright": "0.18.0",
     "eslint-plugin-prettier": "4.2.1",
-    "eslint-plugin-react": "7.32.2",
+    "eslint-plugin-react": "7.33.2",
     "eslint-plugin-react-hooks": "4.6.0",
     "extend": "3.0.2",
     "fs-extra": "11.1.1",
     "hammerjs": "2.0.8",
     "husky": "8.0.3",
-    "lint-staged": "13.2.3",
+    "lint-staged": "13.3.0",
     "mocha-junit-reporter": "2.2.1",
-    "picasso-plugin-hammer": "2.2.4",
-    "picasso-plugin-q": "2.2.4",
-    "picasso.js": "2.2.4",
+    "picasso-plugin-hammer": "2.2.5",
+    "picasso-plugin-q": "2.2.5",
+    "picasso.js": "2.2.5",
     "prettier": "2.8.8",
     "pretty-quick": "3.1.3",
     "prop-types": "^15.8.1",
     "qlik-chart-modules": "0.59.0",
-    "qlik-object-conversion": "0.16.2",
+    "qlik-object-conversion": "0.16.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "scriptappy-from-jsdoc": "0.7.0",
     "yargs": "17.7.2"
   },
   "peerDependencies": {
-    "@nebula.js/stardust": ">=4.2.3"
+    "@nebula.js/stardust": ">=4.6.1"
   }
 }
